requirements.txt
inquirer==3.1.2
openai==0.26.1
python-dotenv==0.21.1
